Move failure mock tests to sibling module
Move the `failure_mocks` test module into `failure_mocks/tests.rs` and
leave the parent file focused on the reusable helper definitions.

Keep the test module using `super::*` so it continues to exercise the
private helper API without changing helper visibility.
